Add testing for PyPi releases
