+++ docker/test/Dockerfile
@@ -6,3 +6,4 @@ RUN apt update && pip install -U pip
 COPY requirements.txt ./
 RUN pip install -r ./requirements.txt
 COPY run_docker_tests.sh ./
+COPY run_pypi_tests.sh ./
